// src/app/business-logic/model/tasks/tasksEditRequest.ts
/**
 * tasks
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 2.13
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { ConfigurationItem } from './configurationItem';
import { Execution } from './execution';
import { Input } from './input';
import { ParamsItem } from './paramsItem';
import { Script } from './script';
import { TaskModels } from './taskModels';
import { TaskTypeEnum } from './taskTypeEnum';
import {Container} from './container';


export interface TasksEditRequest {
    /**
     * ID of the task
     */
    task: string;
    /**
     * If not true, call fails if the task status is not 'created'
     */
    force?: boolean;
    /**
     * Task name Unique within the company.
     */
    name?: string;
    /**
     * User-defined tags list
     */
    tags?: Array<string>;
    /**
     * System tags list. This field is reserved for system use, please don't use it.
     */
    system_tags?: Array<string>;
    type?: TaskTypeEnum;
    /**
     * Free text comment
     */
    comment?: string;
    /**
     * Parent task id Must be a completed task.
     */
    parent?: string;
    /**
     * Project ID of the project to which this task is assigned Must exist[ab]
     */
    project?: string;
    input?: Input;
    /**
     * Output storage id Must be a reference to an existing storage.
     */
    output_dest?: string;
    execution?: Execution;
    script?: Script;
    /**
     * Task hyper params per section
     */
    hyperparams?: { [key: string]: any; };
    /**
     * Task configuration params
     */
    configuration?: { [key: string]: ConfigurationItem; };
    models?: TaskModels;
    container?: Container;
}